Documentation of QCoapClient, QCoapReply, QCoapDiscoveryReply, QCoapRequest, QCoapProtocol and QCoapConnection

+++ src/coap/qcoapinternalreply.cpp
@@ -116,7 +116,7 @@ void QCoapInternalReply::addOption(const QCoapOption& option)
         blockNumber = (blockNumber << 4) | ((optionData[option.length()-1]) >> 4);
         d->currentBlockNumber = blockNumber;
         d->hasNextBlock = ((optionData[option.length()-1] & 0x8) == 0x8);
-        d->blockSize = qPow(2, (optionData[option.length()-1] & 0x7) + 4);
+        d->blockSize = static_cast<uint>(qPow(2, (optionData[option.length()-1] & 0x7) + 4));
     }
 
     d->message.addOption(option);
